remove useless makefile for LLVM/Clang

+++ make_llvm_clang.sh
@@ -1,15 +1,15 @@
 #!/bin/bash
-function print_usage() {
-    echo "Auto download and make the llvm with clang by specific version."
-}
 
-if [ $# != 1 ]
+if [ $# != 2 ]
 then
-    echo 'Please specify the version.'
+    echo 'Usage: $0 <llvm-version> <ycm_cpp_path>'
+    echo 'Example: $0 3.9.1 "~/.vim/plugged/YouCompleteMe/third_party/ycmd/cpp"'
     exit 1
 fi
 
-LLVM_DIR=llvm-src
+CUR_PWD=$(pwd)
+
+LLVM_DIR=$CUR_PWD/llvm-src
 
 BASE_LINK="http://llvm.org/releases/$1/"
 
@@ -23,7 +23,12 @@ CLANG_SRC_TAR_LINK=$BASE_LINK$CLANG_SRC_TAR
 
 COMP_RT_SRC=compiler-rt-$1.src
 COMP_RT_SRC_TAR=$COMP_RT_SRC.tar.xz
-COMP_RT_SRC_TAR=$BASE_LINK$COMP_RT_SRC_TAR
+COMP_RT_SRC_TAR_LINK=$BASE_LINK$COMP_RT_SRC_TAR
+
+echo "LLVM_SRC_TAR LINK: $LLVM_SRC_TAR_LINK"
+echo "LLVM_SRC_TAR LINK: $LLVM_SRC_TAR_LINK"
+echo "CLANG_SRC_TAR_LINK: $CLANG_SRC_TAR_LINK"
+echo "COMP_RT_SRC_TAR_LINK: $COMP_RT_SRC_TAR_LINK"
 
 if [ ! -f $LLVM_SRC_TAR ]; then
     echo "llvm source tar $LLVM_SRC_TAR does not exist!"
@@ -51,7 +56,7 @@ fi
 
 if [ ! -f $COMP_RT_SRC_TAR ]; then
     echo "compiler-rt source tar $COMP_RT_SRC_TAR does not exist!"
-    echo "Begin to donload $COMP_RT_SRC_TAR"
+    echo "Begin to download $COMP_RT_SRC_TAR"
     wget $COMP_RT_SRC_TAR_LINK
     if [ $? != 0]
     then
@@ -61,40 +66,45 @@ if [ ! -f $COMP_RT_SRC_TAR ]; then
     fi
 fi
 
-tar xvf $LLVM_SRC_TAR
-mv $LLVM_SRC $LLVM_DIR
-cd $LLVM_DIR
+if [ ! -f $LLVM_DIR ]; then
+    tar xvf $LLVM_SRC_TAR
+    mv $LLVM_SRC $LLVM_DIR
+    cd $LLVM_DIR
 
-tar xvf ../$CLANG_SRC_TAR -C tools
-mv tools/$CLANG_SRC tools/clang
+    tar xvf ../$CLANG_SRC_TAR -C tools
+    mv tools/$CLANG_SRC tools/clang
 
-tar xvf ../$COMP_RT_SRC_TAR -C projects
-mv projects/$COMP_RT_SRC projects/compiler-rt
+    tar xvf ../$COMP_RT_SRC_TAR -C projects
+    mv projects/$COMP_RT_SRC projects/compiler-rt
+    mkdir build
+fi
 
-mkdir build
-cd build
+cd $LLVM_DIR/build
 
-cmake -G "Unix Makefiles" ..
-if [ $? != 0]; then
+# cmake -G "Unix Makefiles" ..
+# cmake -G "Unix Makefiles" .. -DCMAKE_BUILD_TYPE:"Release" -DLLVM_Targets_to_build:"X86_64"
+cmake -G "Unix Makefiles" .. -DCMAKE_BUILD_TYPE:string=Release -DLLVM_TARGETS_TO_BUILD:string="x86_64"
+
+if [ $? != 0 ]; then
     echo "Failed to cmake Makefiles for llvm and clang."
     exit 1
 fi
 make
-if [ $? != 0]; then
+if [ $? != 0 ]; then
     echo "Failed to make."
+    exit 1
 fi
-
 
 mkdir -p ycm_build
 cd ycm_build
 
-cmake -G "Unix Makefiles" . -DEXTERNAL_LIBCLANG_PATH=../lib/libclang.so ~/.vim/bundle/YouCompleteMe/third_party/ycmd/cpp
+cmake -G "Unix Makefiles" . -DEXTERNAL_LIBCLANG_PATH=../lib/libclang.so $2 
 if [ $? != 0]; then
     echo "Failed to cmake Makefiles!"
 fi
 
 cmake --build . --target ycm_core --config Release
-if [ $? != 0]; then
+if [ $? != 0 ]; then
     echo "Failed to cmake build ycm_core!"
 fi
